requirements/requirements.txt
allure-pytest==2.13.2
pydantic==2.6.0
wiremock==2.6.1